+++ packages/component-library/src/components/plmg-error-message/plmg-error-message.types.ts
@@ -1,0 +1,5 @@
+export const sizes = ['medium', 'large'] as const;
+export type PlmgErrorMessageSize = typeof sizes[number];
+export function isPlmgErrorMessageSize(x: any): x is PlmgErrorMessageSize {
+  return sizes.includes(x);
+}