+++ packages/component-library/src/components/plmg-separator/plmg-separator.stories.js
@@ -47,7 +47,7 @@ export const AllThicknesses = (args) => {
   const htmlContent = thicknesses
     .map(
       (thickness) =>
-        `<div style="width: 300px; height: 300px"><plmg-separator thickness="${thickness}"></plmg separator></div>`
+        `<div style="width: 100px; height: 20px"><plmg-separator thickness="${thickness}"></plmg separator></div>`
     )
     .join('')
     .trim();
@@ -65,7 +65,7 @@ export const AllDirections = (args) => {
   const htmlContent = directions
     .map(
       (direction) =>
-        `<div style="width: 300px; height: 300px;">
+        `<div style="width: 100px; height: 100px;">
             <plmg-separator direction="${direction}"></plmg-separator>
           </div>
         `
@@ -82,11 +82,33 @@ export const AllDirections = (args) => {
 };
 AllDirections.storyName = 'All directions';
 
+export const SomeColors = (args) => {
+  const htmlContent = `
+<div style="width: 100px; height: 20px;">
+<plmg-separator></plmg-separator>
+</div>
+<div style="width: 100px; height: 20px;">
+<plmg-separator color="red"></plmg-separator>
+</div>
+<div style="width: 100px; height: 20px;">
+<plmg-separator color="var(--plmg-color-background-standout)"></plmg-separator>
+</div>
+  `;
+
+  const el = document.createElement('div');
+  el.innerHTML = htmlContent;
+  el.style.display = 'flex';
+  el.style.justifyContent = 'space-between';
+  el.style['flex-wrap'] = 'wrap';
+  return el;
+};
+SomeColors.storyName = 'Some colors';
+
 export const AllVariations = (args) => {
   let htmlContent = '';
   directions.forEach((direction) => {
     thicknesses.forEach((thickness) => {
-      htmlContent += `<div style="width: 300px; height: 300px;">
+      htmlContent += `<div style="width: 100px; height: 100px;">
       <plmg-separator direction="${direction}" thickness="${thickness}"></plmg-separator>
     </div>
   `;